refactor(relay): テストのmockクライアントをmocksモジュールに分離
harness.rsが「クライアント役のmock」「実物の組み立て」「検証ヘルパ」を
全部抱えていたのを、mockをtests/mocks/に切り出して役割を明確化。
harnessは実物data plane部品とmockクライアントの組み立てに専念する。

- mocks/upstream_client.rs: publisher役（ChannelStreamReceiver +
  UpstreamSubgroupStream、オブジェクトビルダー）
- mocks/downstream_client.rs: subscriber役（MockPublisher/MockDataSender、
  EgressEventと記録の問い合わせヘルパ）
