+++ relay/src/modules/relay/tests.rs
@@ -4,3 +4,4 @@
 
 mod data_plane;
 mod harness;
+mod mocks;
